Split Channel and Socket out into their own files
