+++ SwiftPhoenixClient.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,9 @@
 <Workspace
    version = "1.0">
    <FileRef
+      location = "group:README.md">
+   </FileRef>
+   <FileRef
       location = "container:SwiftPhoenixClient.xcodeproj">
    </FileRef>
    <FileRef
